menubar: refactor _create sub-routine: _initializeItems
At this point _create has initialized the
menubar widget and the top level (usually, `li`s)
that define the menubar.

We now want to step through the "actuators"
contained in the menuItems(typically a
`button` or an `a`).

We use an `$.each` here because we need to be
able to apply different behavior wrt the menus
based on whether or not the parent menuItem has
discerned whether it has a subMenu or not.  That
is, we need different behavior if it's a
sub-menu-less menuItem versus a menuItem that
actuates an arbitrary depth sub-menu set.

ALSO:  get rid of non-standard-style helper methods

* `this._applyDOMPropertiesOnItem`
* `this.__applyMouseAndKeyboardBehaviorForMenuItem`
* `this.__applyMouseBehaviorForSubmenuHavingMenuItem`
* `this.__applyKeyboardBehaviorForSubmenuHavingMenuItem`
* `_findNextFocusableTarget`